+++ dfm_tools/__init__.py
@@ -4,4 +4,4 @@
 
 __author__ = """Jelmer Veenstra"""
 __email__ = 'jelmer.veenstra@deltares.nl'
-__version__ = '0.7.49'
+__version__ = '0.7.51'