+++ HISTORY.rst
@@ -1,11 +1,12 @@
 =======
 History
 =======
-0.7.48 (2121-11-01)
+0.7.51 (2121-11-01)
 ------------------
 * increased speed of get_nc.get_netdata() function
 * bugfix for sigma layers cross section and cleaned up cross section definition
-* now possible to retrieve crs_plotdata with get_xzcoords_onintersection() (including hardcoded varname bugfix)
+* now possible to retrieve crs_plotdata with get_xzcoords_onintersection()
+* implemented correct haversine distance calculation for latlon distances
 
 0.7.44 (2121-10-28)
 ------------------